Port qiime.util.which to bipy.app.util.which
As the author of qiime.util.which, I, Jai Ram Rideout, give permission to port
qiime.util.which (GPL code) to the bipy project, which is under the BSD. This
also includes the unit tests.

Note that I modified the code from what it originally was to support an env_var
parameter.

Fixes #45.