Support for documenting private/special methods
Adding numpydoc source so that we can extend/customize as necessary (see licenses/numpydoc.txt for license).

Added an extension to the numpydoc standard that allows a developer to tag a method to be included in the built documentation, when it would have otherwise been ignored. This is necessary in order to have special methods (e.g.  __getitem__, __str__, etc.) included in the built docs.