TST: Add support for NumPy's in-session testing
Tests can now be executed from within a(n) (I)Python session by calling
something like:

>>> import skbio
>>> skbio.test()
.......... .....
XXX Tests OK
>>> skbio.parse.sequences.test()
........ .......

This is as suggested by NumPy's guidelines, see [1] for more information
about this  document for more information about the configuration.

Notes:

The only odd thing is that if we add the few configuration lines as
suggested by Numpy's documentation, then we end up with an error like
this:

>>> import skbio
>>> skbio.test()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/yoshikivazquezbaeza/git_sw/scikit-bio/skbio/__init__.py", line 54, in test
    return Tester().test(level, verbosity)
  File "/Users/yoshikivazquezbaeza/.virtualenvs/skbio/lib/python2.7/site-packages/numpy/testing/nosetester.py", line 146, in __init__
    raise AssertionError
AssertionError
>>>

That's the reason why we are not doing it as the documentation suggests.

[1] https://github.com/numpy/numpy/blob/master/doc/TESTS.rst.txt