MAINT: Remove retrieval of tests and data …
@Jorge-C noted that there's no need to include the filenames of the
tests in the package_data parameter as there's an __init__.py file being
included in each test suite already. As for the test data, given that
the only module that relies on test data is ordination, it is better to
hard-code the path instead of retrieving the information module per
moduel.

Fixes #247