Resolves errors for compare_taxa_summaries.py
This was a real annoying one. The non-parametric tests that the testing for compare_taxa_summaries.py is doing were finding different numbers of values greater than the computed correlation coefficient when using the skbio vs the backports version of correlation_t. Somewhere the machine precision and the type of float or calculation being done is different between the two, even though the number should be exactly the same. I resolved this by rounding to 10 decimal places in the correlation_t non-parametric comparisons.