Added Pearson and Spearman
Added Pearson and Spearman which wrap numpy.corrcoef and scipy.stats.spearmanr (respectively). The wrapping is necessary for now because they calculate things we don't need or the format is wrong for the vast majority of current uses in the code base. This can be rectified at a later time and these wrappers could then go away.