Large commit to add ManWhitneyU
MannWhitneyU and its bootstrapped varients were added in this commit. I made a significant change to the MannWhitneyU code based on what I believe to be the correct interpretation according to scipy documentation. The test code included now makes a lot more sense in terms of agreement between pval betwen the bootstrapped and non-bootstrapped version.