fixed problem with doc string
when we switched over to the scipy one-way ANOVA this doc test p-value started
coming back as nan. it turned out that this was right, because there was no
within-group variance in the groups being compared and our ANOVA wasn't
handling that correctly. this is an example of why we should make a concerted
effort to use scipy when possible for our stats.