Move bioenv code into _bioenv module
Decided to go this route instead of keeping it in the base module as this wasn't a great long-term solution -- we'll be adding several more methods to this subpackage and can't put them all in base! 'nosetests --with-doctest' seems to be happy with this naming scheme and it shouldn't affect users.