ENH: expanded docs, added ignore_comment flag
since parse_fasta is one of the most commonly used functions, expanded the docs
to make it more accessible. also, one of the most common things we'd do with
this was pass a lambda function for label_to_name to strip the comment field
from the label line. added a boolean parameter which will do that for you now,
so we don't have tons of untested lambda functions being passed in all over the
place.