PY3: Force string in HTTPResponse return data
Problem
-------

The codebase relied on the fact that the data being parsed from the HTTP
request was a string. When working with python 3 the result being read
from the response came out as bytes which would break the parsers and
would in turn break the work flow as it currently is laid out.

Solution
--------

Decoding the output as UTF-8 solves the problem, note that this has to
be done in the location where the data is retrieved, otherwise
downstream patches and fixes would need to take place.

Conclusion
----------

This could have been simpler if instead of trying to convert the contents
of an already existant StringIO object we had fixed the creation of the
StringIO itself.

Notes for the reader
--------------------

Think more thoroughly about the changes you are making before making
them.

ASCII Art
---------

                \=/,
                |  @___oo
      /\  /\   / (___,,,}      The Green Dragon
     ) /^\) ^\/ _)
     )   /^\/   _)            "Out of order there is chaos,
     )   _ /  / _)             and out of chaos there is order"
 /\  )/\/ ||  | )_)
<  >      |(,,) )__)
 ||      /    \)___)\
 | \____(      )___) )___
  \______(_______;;; __;;;

Credits
-------

ASCII Art:
http://the.sunnyspot.org/asciiart/gallery/saaas1.html