PY3K/TST: cast range to list and random seeding
Python 3 Related
----------------

The code used a range call to generate a list of elements, and then
access elements from that list. Changing from python 2.7.x to 3.x you
cannot do this as range becomes a generator, thus I've added a cast to a
list.

Testing
-------

Using the built-in seed function from the random library would generate
different values between python 2.7 and py3k. Thus I've changed the
seeding and random operations to use numpy's which is consistent across
python versions.

Note that along with this I had to change the results of the tests,
however these will be consistent now.