Run tests via module invocation
In order to provide proper exit codes, but also to allow testing
without exiting in interactive sessions, tests are now run on the
command line via python -m skbio.test, which invokes the new
skbio/test.py file. Interactively, the skbio.test() method can
still be invoked directly.

As part of this change, the makefile no longer has a verbose
option, but the option is still available in the test method.