MAINT: introducing a change to get the timestamp to update
``checklist.py`` is failing b/c I fixed some typos in
comments in ``_base_cy.pyx``, but those don't result in changes in the
corresponding ``_base_cy.c`` file after re-compiling so the timestamp on the
pyx file is newer than the timestamp on the c file. I'll undo this in the next
commit. Yes, I feel silly doing this.