DOC: remove Python 2/3 guide
The RFC says to modify the Python 2/3 guide to be a guide for helping Python 2
developers become proficient Python 3 developers. Most of the guide would need
to be rewritten, which is a lot of work for docs that aren't that important
compared to other missing scikit-bio docs. Also, there are existing resources
on this topic. We can create an issue if a dev is available to work on a new
guide.