BUG: fix performance regression by reinstating empty metadata optimizations
Fixes #1361 by reinstating the empty/missing metadata optimizations that were removed in #1321.

This is not a commit revert because upstream changes have happened since #1321, causing many merge conflicts. We also want to keep some of the changes in #1321 and discard others, so a manual revert was necessary.

The unit tests testing private state to ensure performance have not been reinstated, as this didn't scale (from a development/codebase perspective) and was fragile to implementation details. Now, private state unit tests exist only for the metadata mixin classes, where that private state is defined. Instead of writing private state unit tests for performance, users and developers are encouraged to submit ASV benchmarks if a method is determined to have suboptimal performance when handling missing/empty metadata.