Added `OrthologTable` best hit function
The input takes in 2 iterables of sequences (only tested for `Protein`).  Creates either a best hit `pd.DataFrame` or all pairwise comparisons as a `pd.DataFrame`.  Any other ideas for the name of the function?  Also, does `skbio` use `joblib` at all? I can make it run in parallel which would be a lot faster for whole genomes.