TST: Ship a pytest runner with every installation (#1633)
Moves the test runner from being a setup command to being a module-level
function. In this patch I also removed the additional configuration
files and make all necessary configurations through the conftest.py file
and the runner's arguments.

All this is necessary so we can properly test skbio upon building it
using conda.