BUG: Fix windows and 32bit support of `unweighted_unifrac`
This forces the tree `child_index` to be exported as 64bit and
also adds a cast (just in case) into `_nodes_by_counts` in the
`_phylogenetic.pyx` cython code.

I have tested that `unweighted_unifrac` succeeds with the change
to the `child_index`. I have not done a full test run, so while
this closes gh-1685, it is very likely that the similar issues
exist elsewhere.

Ideally, it would be great if you find the time to add an Azure job
for windows, or a 32bit job (which will at least find some of these
issues as well).

On a related note: Since this is an `index`, the correct dtype is
likely `np.intp`, which is 32bit on 32bit systems and 64bit on
64 bit systems. So it might be better if the cython `DTYPE_t` was
changed that (and the `child_index` code here as well), but
of course that relies on `DTYPE_t` not being used for anything
that is not an index operation but should have 64bit accuracy.